Moved the concepts into their own folders
