+++ Plugins/UnrealRanges/Source/UnrealRanges/Public/Ranges/Concepts/UEContainer.h
@@ -3,9 +3,9 @@
 #pragma once
 
 #include "CoreMinimal.h"
+#include <range/v3/range/concepts.hpp>
 
 namespace UE::Ranges {
-
     namespace Detail {
 
         /**
